+++ api/handlers/utils.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/jmpsec/osctrl/pkg/logging"
-	"github.com/jmpsec/osctrl/types"
+	"github.com/jmpsec/osctrl/pkg/types"
 	"github.com/jmpsec/osctrl/utils"
 	"github.com/rs/zerolog/log"
 	"gorm.io/gorm"
