+++ carves/s3.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/jmpsec/osctrl/pkg/settings"
-	"github.com/jmpsec/osctrl/types"
+	"github.com/jmpsec/osctrl/pkg/types"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/viper"
