+++ cli/api-environment.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/jmpsec/osctrl/pkg/environments"
 	"github.com/jmpsec/osctrl/pkg/settings"
-	"github.com/jmpsec/osctrl/types"
+	"github.com/jmpsec/osctrl/pkg/types"
 )
 
 // GetEnvironments to retrieve all environments from osctrl
