+++ pkg/logging/splunk.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/jmpsec/osctrl/pkg/settings"
-	"github.com/jmpsec/osctrl/types"
+	"github.com/jmpsec/osctrl/pkg/types"
 	"github.com/jmpsec/osctrl/utils"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/viper"
