+++ pkg/types/osquery.go
