+++ tls/handlers/carves.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"github.com/jmpsec/osctrl/carves"
 	"github.com/jmpsec/osctrl/pkg/settings"
-	"github.com/jmpsec/osctrl/types"
+	"github.com/jmpsec/osctrl/pkg/types"
 	"github.com/rs/zerolog/log"
 )
