+++ users/users.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/golang-jwt/jwt/v4"
-	"github.com/jmpsec/osctrl/types"
+	"github.com/jmpsec/osctrl/pkg/types"
 	"github.com/jmpsec/osctrl/utils"
 	"github.com/rs/zerolog/log"
 	"golang.org/x/crypto/bcrypt"
